Don't play the video in webgl on Android FF
Currently it is not possible to play videos in a webgl context
due to being it a bug in the browser. This commit adds a small bit
of detection code to disable playing the video in the webgl cuboid.

This code "should" automatically enable the rendering when it gets
fixed. But in the meantime, keep an eye on the bugtracker:
https://bugzilla.mozilla.org/show_bug.cgi?id=1699486